Organize appmanager package
Problem: appManager was becoming a mess of methods and files were becoming very large.

Solution: Split common methods into separate files for better organization. Put struct types into
a types.go, per the golang convention.
