+++ pkg/appmanager/routing.go
@@ -18,6 +18,7 @@ package appmanager
 
 import (
 	"bytes"
+	"fmt"
 	"net/url"
 	"sort"
 	"strconv"
@@ -28,6 +29,8 @@ import (
 
 	"k8s.io/client-go/pkg/apis/extensions/v1beta1"
 )
+
+const httpRedirectRuleName = "http-redirect"
 
 func (r Rules) Len() int           { return len(r) }
 func (r Rules) Less(i, j int) bool { return r[i].FullURI < r[j].FullURI }
@@ -178,4 +181,21 @@ func processIngressRules(
 
 	rls = append(rls, w...)
 	return &rls
+}
+
+func newHttpRedirectPolicyRule(httpsPort int32) *Rule {
+	loc := fmt.Sprintf(`tcl:https://[getfield [HTTP::host] ":" 1]:%d[HTTP::uri]`, httpsPort)
+	redirAction := action{
+		Name:      "0",
+		HttpReply: true,
+		Location:  loc,
+		Redirect:  true,
+		Request:   true,
+	}
+	rule := Rule{
+		Name:    httpRedirectRuleName,
+		Actions: []*action{&redirAction},
+		Ordinal: 0,
+	}
+	return &rule
 }
